fix: empty groomed sequences break per-jet inference on real data
`nn.GRU` raises "Expected sequence length to be larger than 0" on a zero-length
sequence, so any jet with an EMPTY hadron-level tree crashed every per-jet
inference path — the closure loop, the calibration loops, TARP, serving.

This is not a corner case. In `cpp/test_data/jets.root` (PYTHIA 8.3, z_cut=0.1,
k_t floor 1 GeV) 6.9% of jets have no hadron-level primary emission and 16.0% have
no parton-level one. It went unnoticed because the synthetic generator produces
neither, and because batched training never sees it: `collate` pads to the batch
maximum, so Mx >= 1 as long as one jet in the batch is non-empty. Only the per-jet
routes, where Mx = nx, hit it. Found by running the new calibration notebook on
the real sample.

- `GRUEncoder._states` short-circuits Mx == 0 to zero states. The empty sequence
  then pools to zeros and `e(x)` is the projection's bias — the encoder's "I was
  told nothing" context. `lundnet` and `deepsets` already handled it.
- `ARJunipr.xattn_kv` returns None when there are no hadron nodes at all, and
  `_apply_xattn` drops the residual for rows whose keys are entirely masked.
  Softmax over nothing is undefined; torch 2.11 happens to return zeros rather
  than NaN, but that is not promised, and a NaN there would silently poison a
  whole batch's gradients. Dropping the residual is also the right semantics.

`tests/test_empty_sequences.py` covers both empty levels across all six families
and all three encoders, and pins that the batched and per-jet routes agree for an
empty jet — otherwise training and evaluation would model different things.

+++ src/h2p_rsd_junipr/models/cfm.py
@@ -177,7 +177,7 @@ class CFM(PosteriorModel):
         `u_t = x_1 - (1 - sigma_min) x_0`, constant along the path."""
         x1 = s
         x0 = torch.randn_like(x1)
-        t = torch.rand(x1.shape[:-1] + (1,), device=x1.device)
+        t = torch.rand(x1.shape[:-1] + (1,), device=x1.device, dtype=x1.dtype)
         a = 1.0 - (1.0 - self.sigma_min) * t
         x_t = a * x0 + t * x1
         target = x1 - (1.0 - self.sigma_min) * x0
@@ -191,7 +191,7 @@ class CFM(PosteriorModel):
         the Hutchinson estimator every high-dimensional CNF is forced into is
         unnecessary here, and the likelihood comes out deterministic rather than
         stochastic (the property that makes it usable for model selection)."""
-        div = torch.zeros(xx.shape[:-1], device=xx.device)
+        div = torch.zeros(xx.shape[:-1], device=xx.device, dtype=xx.dtype)
         for i in range(4):
             e = torch.zeros_like(v)
             e[..., i] = 1.0
@@ -215,9 +215,13 @@ class CFM(PosteriorModel):
         n_steps = int(n_steps or self.n_ode_steps)
         solver = self.ode_solver
         dt = (-1.0 if reverse else 1.0) / n_steps
-        t = (torch.ones if reverse else torch.zeros)(s.shape[:-1] + (1,), device=s.device)
+        # dtype from `s`, not the default: the likelihood is evaluated in float64 by the
+        # normalization test, and a float32 time would only survive by type promotion.
+        t = (torch.ones if reverse else torch.zeros)(
+            s.shape[:-1] + (1,), device=s.device, dtype=s.dtype
+        )
         x = s
-        acc = torch.zeros(s.shape[:-1], device=s.device) if with_divergence else None
+        acc = torch.zeros(s.shape[:-1], device=s.device, dtype=s.dtype) if with_divergence else None
 
         def f(xx, tt):
             if not with_divergence:
